probe: performance enhancements
- Run out of SRAM, it goes quite a lot faster
- Cache the DAP clock frequency
- Read RDATA and write WDATA all in one go, and merge the turnaround
  into the ACK phase. Reduces the number of missed PIO SM cycles during an
  SWD transfer

Also fix a compiler warning and demote the loglevel of some debug messages.

+++ src/probe.h
@@ -27,8 +27,8 @@
 #define PROBE_H_
 
 void probe_set_swclk_freq(uint freq_khz);
-void probe_write_bits(uint bit_count, uint8_t data_byte);
-uint8_t probe_read_bits(uint bit_count);
+void probe_write_bits(uint bit_count, uint32_t data_byte);
+uint32_t probe_read_bits(uint bit_count);
 
 void probe_read_mode(void);
 void probe_write_mode(void);
